Moved synch_to_reference to multistream
